Refactor admin attendees route into focused sub-files
Split src/routes/admin/attendees.ts (~1150 lines) into:
- attendees-utils.ts: shared helpers (attendeeGetRoute, verifiedAttendeeForm, getReturnUrl, NO_PROVIDER_ERROR, etc.)
- attendees-merge.ts: merge attendee route handlers (GET + POST)
- attendees-edit.ts: edit page, event links, and refresh-payment handlers
- attendees.ts: trimmed coordinator (delete, checkin, add, resend + route table)

Updated attendee-refunds.ts to import shared utilities from attendees-utils.ts directly.
